エディタのクリア時にモーダルウィンドウで確認するように変更

+++ src/components/EditArea.vue
@@ -1,15 +1,17 @@
 <template>
     <div id="editArea">
+        <ClearMessage ref="clearMessage" @confirmed-clear-text="confirmedClearText"/>
+
         <el-row>
             <Editor ref="editor"/>
         </el-row>
         <el-row type="flex" justify="space-around">
             <el-col :span="6"></el-col>
             <el-col :span="6">
-                <ClearButton @clear-text="clearText"/>
+                <ClearButton @pushed-clear-button="pushedClearButton"/>
             </el-col>
             <el-col :span="6">
-                <ExecuteButton @execute-program="executeProgram"/>
+                <ExecuteButton @pushed-execute-button="pushedExecuteButton"/>
             </el-col>
             <el-col :span="6"></el-col>
         </el-row>
@@ -21,6 +23,7 @@
     import Editor from './Editor'
     import ClearButton from './ClearButton'
     import ExecuteButton from './ExecuteButton'
+    import ClearMessage from './ClearMessage'
 
     export default {
         data () {
@@ -32,14 +35,18 @@
             'Editor': Editor,
             'ClearButton': ClearButton,
             'ExecuteButton': ExecuteButton,
+            'ClearMessage': ClearMessage
         },
         methods: {
-            clearText: function(){
-                this.$refs.editor.clear();
+            pushedClearButton: function(){
+                this.$refs.clearMessage.confirm()
             },
-            executeProgram: function(){
+            pushedExecuteButton: function(){
                 const code = this.$refs.editor.getText();
                 this.$emit('update-text', code);
+            },
+            confirmedClearText: function(){
+                this.$refs.editor.clear()
             }
         }
     }
